sched/eevdf: avoid pick_eevdf() returns NULL
pick_eevdf() may return NULL, which would triggers NULL pointer
dereference and crash when best and curr are both NULL.

There are two cases when curr would be NULL:
	1) curr is NULL when enter pick_eevdf
	2) we set it to NUll when curr is not on_rq or eligible.

And when we went to the best = curr flow, the se should never be NULL,
So when best and curr are both NULL, we'd better set best = se to avoid
return NULL.

Below crash is what I encounter very low probability on our server and
I have not reproduce it, and I also found other people feedback some
similar crash on lore. So believe the issue is really exit.

<1>[    8.607396] Unable to handle kernel NULL pointer dereference at virtual address 00000000000000a0
<1>[    8.607399] Mem abort info:
<1>[    8.607400]   ESR = 0x0000000096000004
<1>[    8.607401]   EC = 0x25: DABT (current EL), IL = 32 bits
<1>[    8.607402]   SET = 0, FnV = 0
<1>[    8.607403]   EA = 0, S1PTW = 0
<1>[    8.607403]   FSC = 0x04: level 0 translation fault
<1>[    8.607404] Data abort info:
<1>[    8.607404]   ISV = 0, ISS = 0x00000004
<1>[    8.607404]   CM = 0, WnR = 0
<1>[    8.607405] user pgtable: 4k pages, 48-bit VAs, pgdp=000000011efef000

Signed-off-by: limingming3 <limingming3@lixiang.com>