android: binder: fix UAF in binder_thread_release
When a thread is being released via BINDER_THREAD_EXIT, it's possible
for another thread to obtain a pointer to the thread object before it's
fully destroyed, leading to a use-after-free condition.

Call trace:
binder_ioctl+0x1a8/0x110c
__se_sys_ioctl+0xd0/0x114
__arm64_sys_ioctl+0x1c/0x24

Unable to handle kernel paging request at virtual address 000080000000002a

This occurs because binder_get_thread() can return a thread that is in the
process of being destroyed, but hasn't been removed from the process's
thread list yet.

Fix it by:
1. Adding a "dead" flag to struct binder_thread.
2. Setting the flag in binder_thread_release() under proc lock.
3. Checking the flag in binder_get_thread() with READ_ONCE().
4. Initializing the flag to false when a new thread is created.

This prevents use-after-free by ensuring that any thread marked as dead
is treated as invalid and not returned by binder_get_thread().

Signed-off-by: reigadegr <2722688642@qq.com>